+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # v5.2.0 (latest)
+###  New features
  - delete reminders for form actions when the correct form is filled
  - add `email_activate_subject` and `email_reminder_subject` fields for configuration when sending emails
  - add an option to attach files when schedule a mail action for forms
  - add language check `condition`
+ - add an option for multiple selection of jobs in `scheduledJobs` and then `execute` them or `delete` them
 
 # v5.1.5
 ### Bugfix